The great dictionary purge
change game_info to a class. Update info_for_player to use the class
and fix tests until all pass.
run tournament and sample scripts to test existing bot

The README does not reference the old dictionary style, which was
either a bug or a feature. It was a convenience either way

I'm intending on adding helper methods to GameInfo, but keeping
the first attempt just a refactor